+++ Utils/azure-explorer-common/src/com/microsoft/tooling/msservices/serviceexplorer/azure/webapp/deploymentslot/DeploymentSlotModule.java
@@ -35,8 +35,9 @@ public class DeploymentSlotModule extends AzureRefreshableNode implements Deploy
     }
 
     @Override
-    public void renderDeploymentSlots(final String webAppId, @NotNull final List<DeploymentSlot> slots) {
-        slots.forEach(slot -> addChildNode(new DeploymentSlotNode(slot.id(), webAppId,this, slot.name(), slot.state(),
-            slot.operatingSystem().toString(), this.subscriptionId, slot.defaultHostName())));
+    public void renderDeploymentSlots(@NotNull final List<DeploymentSlot> slots) {
+        slots.forEach(slot -> addChildNode(
+            new DeploymentSlotNode(slot.id(), slot.parent().id(), this, slot.name(), slot.state(),
+                slot.operatingSystem().toString(), this.subscriptionId, slot.defaultHostName())));
     }
 }
