+++ Utils/azure-explorer-common/src/com/microsoft/tooling/msservices/serviceexplorer/azure/webapp/deploymentslot/DeploymentSlotModulePresenter.java
@@ -8,9 +8,7 @@ public class DeploymentSlotModulePresenter<V extends DeploymentSlotModuleView> e
     public void onRefreshDeploymentSlotModule(final String subscriptionId, final String webAppId) throws IOException {
         final DeploymentSlotModuleView view = getMvpView();
         if (view != null) {
-            view.renderDeploymentSlots(
-                webAppId,
-                AzureWebAppMvpModel.getInstance().getDeploymentSlots(subscriptionId, webAppId));
+            view.renderDeploymentSlots(AzureWebAppMvpModel.getInstance().getDeploymentSlots(subscriptionId, webAppId));
         }
     }
 }
